Split machines.h into different headers
