+++ LUINT/machine.cpp
@@ -183,72 +183,4 @@ namespace LUINT::Machines
 	}
 
 #pragma endregion Rendering
-
-	StateMachine::StateMachine(Data::SessionData& _session, std::string _name, Network* _network) : Machine(_session, _name, _network)
-	{
-		state = luaL_newstate();
-	}
-
-	Monitor::Monitor(Data::SessionData& _session, std::string _name, Network* _network) : Machine(_session, _name, _network)
-	{
-	}
-
-	void thing()
-	{
-		std::cout << "works";
-	}
-
-	void LED::ImplementLua(lua_State * state, sol::table&& proxy_table)
-	{
-		sol::state_view lua(state);
-
-		proxy_table.set_function("set_state", &LED::f_set_state, this); //Why doesn't this compile???
-		proxy_table.set_function("get_state", &LED::f_get_state, this);
-	}
-
-	void LED::RenderWindow()
-	{
-		ImGui::Text(turnedOn ? "On" : "Off");
-	}
-
-	void Keyboard::RenderWindow()
-	{
-		//ImGui::TextUnformatted("Type here!");
-		//ImGuiIO& io = ImGui::GetIO();
-
-		//for (int i = 0; i < 512; i++)
-		//{
-		//	if (io.KeysDownDuration[i] == 0.f)
-		//	{
-		//		// Just pressed key!
-		//		for (auto& connection : session->connections)
-		//		{
-		//			if (connection.first == this)
-		//			{
-
-		//			}
-		//			if (connection.second == this)
-		//			{
-		//				
-		//			}
-		//		}
-		//	}
-		//}
-	}
-
-	void Button::RenderWindow()
-	{
-		ImGui::Selectable("Button");
-		if (ImGui::IsItemClicked(0))
-		{
-			network->SendEvent("button_pressed", uid);
-			pressed = true;
-		}
-		if (ImGui::IsMouseReleased(0) && pressed)
-		{
-			network->SendEvent("button_released", uid);
-			pressed = false;
-		}
-		ImGui::SetWindowSize(ImGui::GetContentRegionAvail(), ImGuiCond_Appearing);
-	}
 }
